==> Extended_Figure8/02_NMF_CD4_T_cell_Projection/00_Download_NMFprojection_Matrix.ipynb <==
import requests
import gzip
import shutil

url = 'https://github.com/yyoshiaki/NMFprojection/raw/main/data/NMF.W.CD4T.csv.gz'
gz_filename = 'NMF.W.CD4T.csv.gz'
csv_filename = 'NMF.W.CD4T.csv'

# Download the file
response = requests.get(url, stream=True)
if response.status_code == 200:
    with open(gz_filename, 'wb') as file:
        file.write(response.content)
    print(f"Downloaded '{gz_filename}'.")
else:
    print(f"Failed to download file. Status code: {response.status_code}")

# Unzip the .gz file
with gzip.open(gz_filename, 'rb') as f_in:
    with open(csv_filename, 'wb') as f_out:
        shutil.copyfileobj(f_in, f_out)
    print(f"Unzipped '{gz_filename}' to '{csv_filename}'.")
