==> Extended_Figure8/02_NMF_CD4_T_cell_Projection/08_Extract_Th2_and_Tfh_Baseline_Samples_BRI.ipynb <==
suppressPackageStartupMessages(library(plyr))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(tibble))
library(parallel)

meta_data=read.csv("/home/jupyter/BRI_Figures_Final_V1/Dataset/scRNA_meta_data-2024-05-09.csv")
meta_data_subset<-meta_data %>% 
filter(Covid_exclusion=='no',sample.daysSinceFirstVisit<600) %>%
filter(!sample.visitName%in% c('Flu Year 1 Day 7',
                              'Flu Year 2 Day 7'))

df_list<-mclapply(meta_data_subset$pbmc_sample_id,function(x){
df<-read.csv(paste0("/home/jupyter/BRI_Figures_Final_V1/Figure6/02_NMF_CD4_T_cell_Projection/NMF_Score_BRI/",x,".csv"))
df$pbmc_sample_id=x
return(df)
},mc.cores=60)

df_combined<-do.call(rbind,df_list)
df_combined_longformat<- pivot_longer(df_combined, cols = starts_with("NMF"), names_to = "NMF_Scores", values_to = "value")

df_combined_longformat<-df_combined_longformat %>% group_by(pbmc_sample_id,NMF_Scores,AIFI_L3) %>% summarise(AverageScore=mean(value))

df_combined_longformat_filtered<-df_combined_longformat %>% filter(NMF_Scores %in% c('NMF5_Th2','NMF6_Tfh')) %>% filter(AIFI_L3 %in% c('CM CD4 T cell'))

df<-left_join(df_combined_longformat_filtered,meta_data_subset)

write.csv(df,'Th2_Tfh_Baseline_Samples_BRI.csv')
